Fix memory leaks in palindrome.c and main.c
This commit fixes the memory leaks present in palindrome.c and main.c. The memory leak in main resulted from it calling palindrome, which allocates memory for the final result it generates, but this memory is never deallocated. So, we created a char pointer called answer, and then stored the final result of palindrome in that variable. We then freed that variable once we are done with it. In palidrome.c, the memory leak resulted from the function str_reverse allocating memory for its final result, but this memory is never deallocated. the function palindrome then calls srt_reverse and stores it in rev. We found the last place where the information in rev was used and then added a free statement after that.

Co-authored-by: Dante Miller <mill8622@morris.umn.edu>