Fix memory leaks in palindrome_test.cpp
This commit fixes the memory leaks present in the palindrome_test.cpp file. The memory leaks resulted from the functions is_palindrome and not_palindrome calling palindrome(str) and then never deallocating the memory that is allocated during that function. So, to fix it, in both functions we created a char pointer called answer, stored the result of palindrome in answer, did ASSERT_STREQ(answer, "Yes") and ASSERT_STREQ(answer, "No") in their respective functions , and then freed the memory stored in answer after the assert statement.

Co-authored-by: Dante Miller <mill8622@morris.umn.edu>